backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
pydantic[email]==2.5.3
google-generativeai==0.3.2
python-multipart==0.0.6
